Imported component scripts into index.js and layout.pug
