+++ public/javascripts/index.js
@@ -1,17 +1,8 @@
-const toggleDropdown = () => {
-  const button = document.querySelector(".dropdown-button");
-  button.addEventListener("click", (event) => {
-    const dropdown = document.querySelector(".dropdown-container");
-
-    if (dropdown.style.display === "none") {
-      dropdown.style.display = "flex";
-    } else {
-      dropdown.style.display = "none";
-    }
-  });
-};
+import { toggleDropdown } from "./dropdown.js";
+import { follow } from "./follow-button.js";
 
 window.addEventListener("load", (event) => {
   console.log("hello from javascript!");
   toggleDropdown();
+  follow();
 });
