+++ views/layout.pug
@@ -6,6 +6,8 @@ html
         link(rel='stylesheet' href='/stylesheets/main-styles.css')
         link(rel="stylesheet" href="/stylesheets/components/main-nav.css")
         link(rel="stylesheet" href="/stylesheets/components/dropdown.css")
+
+        script(src="/javascripts/index.js" type="module" defer)
   body
     if locals.authenticated
       include main-nav.pug
